requirements.txt
streamlit
streamlit-js-eval
google-genai
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
Pillow
requests
